requirements.txt
requests
pandas
python-dotenv
reportlab
yagmail
schedule
apscheduler
